requirements.txt
aniso8601==1.2.1
click==6.7
Flask==0.12.2
Flask-RESTful==0.3.6
gunicorn==19.7.1
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
pyswisseph==2.0.0.post2
python-dateutil==2.6.1
pytz==2017.2
six==1.10.0
Werkzeug==0.12.2
